+++ R/tutorial.R
@@ -1,5 +1,0 @@
-
-tutorialLaunchBrowser <- function(url) {
-  callFun("tutorialLaunchBrowser", url)
-}
-